analyse_marketing/requirements.txt
pandas==2.0.3
numpy==1.24.4
tushare>=1.2
tqdm>=4.66
pyyaml>=6.0
python-dateutil>=2.9
pyarrow==12.0.1
akshare>=1.11.75

